Allow popping based on token at indented line
For example, with

    foo = [x | let a = 3
      ] -- <-- indenting this line

when indenting the second line, the local declarations cannot continue
since the list is done due to the right bracket. Therefore the set of
potential indentations can be narrowed down, by popping the list parser.

Also simplify the code.

Removed "s:Or" since its semantics were unclear, in particular: Should
it retry with the next alternative after a parse failed halfway through?
It was only used by "s:Opt" anyway.