Fix <Plug>HaskellIndentN/-P on Neovim
In Neovim ^F causes indent to happen *twice*. Therefore on the second
round the indentation did just get reset to the default.

This commit works around that by skipping the second indent when we
expect it.

A more robust solution could be something like

    <C-\><C-O>:set indentexpr=<indent><CR><C-F>

which would work regardless of Neovim behavior. == can not be used for
blank lines, and :left would have to be followed by ^F anyway to set the
`did_ai` flag, which is needed in order to act similarly to <C-T>/<C-D>
(that leave `did_ai` as is) where indent on blank lines is deleted upon
leaving Insert mode.

See neovim/neovim#12894